Add Entities.py module
Refactored coin and monster-specific logic into a new module, Entities.py, in the actors folder.
