+++ src/main.py
@@ -6,7 +6,7 @@ import sys
 import path_utils as pu
 from datetime import timedelta
 from game_state.Interface import UI, Screen
-from actors.DrawObject import RainingItem
+from actors.Entities import Coin, Monster
 from actors.Player import Robot
 
 class RainingCoins:
@@ -157,35 +157,6 @@ class GameState:
     
     def is_game_over(self):
         return self.__game_over
-
-class Coin(RainingItem):
-    def __init__(self, game_instance: RainingCoins, coin: pg.image):
-        super().__init__(game_instance, coin)
-
-    def _start_positions(self):
-        return [self._randomize_pos() for i in range(self._frequency)]
-
-    def _collide_condition(self, i: int):
-        self._positions[i][0], self._positions[i][1] = self._randomize_pos()
-        self._count += 1
-
-class Monster(RainingItem):
-    def __init__(self, game_instance: RainingCoins, monster: pg.image):
-        super().__init__(game_instance, monster)
-        self.__visited = [False] * self._frequency
-
-    def _collide_condition(self, i: int):
-        self._game_instance.state.game_over()
-
-    def _monster_count(self, i: int):
-        if (not self.__visited[i] and self._positions[i][1] >= 0 and \
-        self._positions[i][1] + self._image.get_height() <= self._game_instance.screen.height):
-            self._count += 1
-            self.__visited[i] = True
-
-    def _floor_boundary(self, current_monster: list[int], i: int):
-        current_monster[0], current_monster[1] = self._randomize_pos()
-        self.__visited[i] = False
 
 ###########################################################################################################################
